resumate_backend/requirements.txt
fastapi==0.110.0
uvicorn==0.30.0
pydantic==2.6.3
pydantic-settings==2.2.1
sqlalchemy==2.0.29
python-dotenv==1.0.1
cerebras_cloud_sdk==1.29.0